+++ rv/modules/drumsynth.py
@@ -4,6 +4,10 @@ from rv.controller import Controller
 from rv.modules import Behavior as B, Module
 
 
+def panning_controller():
+    return Controller((-128, 128), 0)
+
+
 class DrumSynth(Module):
 
     name = mtype = "DrumSynth"
@@ -13,7 +17,7 @@ class DrumSynth(Module):
     behaviors = {B.receives_notes, B.sends_audio}
 
     volume = Controller((0, 512), 256)
-    panning = Controller((-128, 128), 0)
+    panning = panning_controller()
     polyphony_ch = Controller((1, 8), 4)
     bass_volume = Controller((0, 512), 200)
     bass_power = Controller((0, 256), 256)
@@ -24,6 +28,9 @@ class DrumSynth(Module):
     snare_volume = Controller((0, 512), 256)
     snare_tone = Controller((0, 256), 128)
     snare_length = Controller((0, 256), 64)
+    bass_panning = panning_controller()
+    hihat_panning = panning_controller()
+    snare_panning = panning_controller()
 
 
 class DRUMNOTE(IntEnum):